4-22演習課題　修正1

+++ 4-22/src/Main4/Employee.java
@@ -3,9 +3,9 @@ package Main4;
 public class Employee {
     private String name;
     private int age;
-    private int salary;
+    private double salary;
 
     public String getName() { return name; }
     public int getAge() { return age; }
-    public int getSalary() { return salary; }
+    public double getSalary() { return salary; }
 }
